requirements.txt
accelerate==0.23.0
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.0.0
anytree==2.9.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bert-score==0.3.13
bleach==6.1.0
blis==0.7.11
boto3==1.28.62
botocore==1.31.62
bpemb==0.3.4
catalogue==1.0.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.0.4
cloudpickle==2.2.1
comm==0.1.4
conllu==4.5.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
docopt==0.6.2
editdistance==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.12.4
flair==0.12.2
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gdown==4.4.0
gensim==3.8.3
huggingface-hub==0.17.3
hydra-core==1.3.2
hyperopt==0.2.7
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
Janome==0.5.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.4.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
langdetect==1.0.9
language-tool-python==2.7.1
lemminflect==0.2.3
lru-dict==1.2.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==10.1.0
mpld3==0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
notebook==7.0.4
notebook_shim==0.2.3
num2words==0.5.12
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
OpenHowNet==2.0
overrides==7.4.0
packaging==23.2
pandarallel==1.6.5
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pinyin==0.4.0
plac==1.1.3
platformdirs==3.11.0
plotly==5.17.0
pptree==3.1
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==13.0.0
pycld2==0.41
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch_revgrad==0.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.4
s3transfer==0.7.0
sacremoses==0.0.53
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
segtok==1.5.11
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==2.3.9
sqlitedict==2.1.0
srsly==1.0.7
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
terminaltables==3.1.10
textattack==0.3.9
texthero==1.1.0
thinc==7.4.6
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.14.1
tomli==2.0.1
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformer-smaller-training-vocab==0.3.2
transformers==4.34.0
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
uri-template==1.3.0
urllib3==1.26.17
wasabi==0.10.1
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
Wikipedia-API==0.6.0
word2number==1.1
wordcloud==1.9.2
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
